feat(vue3): rename directives folder to BIMDataDirectives
