+++ src/BIMDataComponents/BIMDataSelect/BIMDataSelect.vue
@@ -55,7 +55,7 @@
 </template>
 
 <script>
-import clickaway from "../../directives/click-away";
+import clickaway from "../../BIMDataDirectives/click-away";
 
 import BIMDataIcon from "../BIMDataIcon/BIMDataIcon.vue";
 import BIMDataCheckbox from "../BIMDataCheckbox/BIMDataCheckbox.vue";
