+++ src/BIMDataDirectives/copy.js
