+++ src/web/views/Elements/Code/Code.vue
@@ -11,8 +11,8 @@
 </template>
 
 <script>
-import highlight from "../../../../directives/highlight.js";
-import copy from "../../../../directives/copy.js";
+import highlight from "../../../../BIMDataDirectives/highlight.js";
+import copy from "../../../../BIMDataDirectives/copy.js";
 import Alerts from "../Alerts/Alerts.vue";
 
 export default {
